+++ src/assets/CMakeLists.txt
