+++ src/assets/importers/font_importer.cpp
@@ -1,4 +1,4 @@
-#include "asset_management/importers/font_importer.hpp"
+#include "assets/importers/font_importer.hpp"
 
 #include "graphics/font.hpp"
