+++ src/assets/importers/material_importer.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "asset_management/assets_fwd.hpp"
+#include "assets/assets_fwd.hpp"
 
-#include "asset_management/type_importer.hpp"
+#include "assets/type_importer.hpp"
 
 class material_importer : public assets::type_importer<graphics::material>
 {
