+++ src/assets/importers/model_importer.hpp
@@ -2,7 +2,7 @@
 
 #include "graphics/graphics_fwd.hpp"
 
-#include "asset_management/type_importer.hpp"
+#include "assets/type_importer.hpp"
 
 class model_importer : public assets::type_importer<graphics::mesh>
 {
