+++ src/assets/importers/texture_importer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "asset_management/type_importer.hpp"
+#include "assets/type_importer.hpp"
 
 class texture_importer : public assets::type_importer<graphics::texture>
 {
@@ -8,5 +8,6 @@ public:
     static constexpr std::string_view extensions { ".png,.jpg,.jpeg" };
     void internal_load(common::file& asset_file) override;
 
-    void internal_update(std::shared_ptr<graphics::texture>, common::file&) override;
+    void internal_update(std::shared_ptr<graphics::texture>,
+                         common::file&) override;
 };
