+++ src/assets/type_importer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "asset_management/asset_manager.hpp"
+#include "assets/asset_manager.hpp"
 #include "common/filesystem.hpp"
 #include "common/logging.hpp"
