+++ src/samples/multiple_viewports.cpp
@@ -2,7 +2,7 @@
 #include <GLFW/glfw3.h>
 /* clang-format on */
 
-#include "asset_management/asset_manager.hpp"
+#include "assets/asset_manager.hpp"
 #include "common/filesystem.hpp"
 #include "common/logging.hpp"
 #include "core/viewport.hpp"
